+++ .changeset/eighty-spies-beam.md
@@ -1,0 +1,5 @@
+---
+'@emotion/css': minor
+---
+
+Source code has been migrated to TypeScript. From now on type declarations will be emitted based on that, instead of being hand-written.
