+++ .changeset/modern-tomatoes-work.md
@@ -1,0 +1,5 @@
+---
+'@emotion/css': patch
+---
+
+Fixed `options` parameter to `createEmotion` from `@emotion/css/create-instance` incorrectly being marked as optional when it's required.
