+++ packages/css/create-instance/package.json
@@ -2,7 +2,7 @@
   "main": "dist/emotion-css-create-instance.cjs.js",
   "module": "dist/emotion-css-create-instance.esm.js",
   "umd:main": "dist/emotion-css-create-instance.umd.min.js",
-  "types": "../types/create-instance",
+  "types": "dist/emotion-css-create-instance.cjs.d.ts",
   "preconstruct": {
     "umdName": "createEmotion"
   }
