+++ packages/css/package.json
@@ -4,7 +4,7 @@
   "description": "The Next Generation of CSS-in-JS.",
   "main": "dist/emotion-css.cjs.js",
   "module": "dist/emotion-css.esm.js",
-  "types": "types/index.d.ts",
+  "types": "dist/emotion-css.cjs.d.ts",
   "files": [
     "src",
     "dist",
@@ -53,8 +53,8 @@
   "preconstruct": {
     "umdName": "emotion",
     "entrypoints": [
-      "./index.js",
-      "./create-instance.js"
+      "./index.ts",
+      "./create-instance.ts"
     ]
   }
 }
