+++ packages/css/src/create-instance.ts
@@ -1,24 +1,80 @@
 import createCache from '@emotion/cache'
-import { serializeStyles } from '@emotion/serialize'
+import {
+  serializeStyles,
+  CSSInterpolation,
+  Interpolation
+} from '@emotion/serialize'
 import {
   insertStyles,
-  getRegisteredStyles
-  /* type EmotionCache, */
-  /* type SerializedStyles */
+  getRegisteredStyles,
+  SerializedStyles,
+  RegisteredCache
 } from '@emotion/utils'
+import { EmotionCache, Options } from '@emotion/cache'
+import { StyleSheet } from '@emotion/sheet'
 
-function insertWithoutScoping(cache, serialized /*: SerializedStyles */) {
+export type {
+  CSSInterpolation,
+  ArrayCSSInterpolation,
+  ComponentSelector,
+  CSSObject
+} from '@emotion/serialize'
+
+function insertWithoutScoping(
+  cache: EmotionCache,
+  serialized: SerializedStyles
+) {
   if (cache.inserted[serialized.name] === undefined) {
     return cache.insert('', serialized, cache.sheet, true)
   }
 }
 
+export type { EmotionCache, Options }
+
+export interface ArrayClassNamesArg extends Array<ClassNamesArg> {}
+export type ClassNamesArg =
+  | undefined
+  | null
+  | string
+  | boolean
+  | { [className: string]: boolean | null | undefined }
+  | ArrayClassNamesArg
+
+export interface CSSStyleSheet extends StyleSheet {
+  speedy(value: boolean): void
+}
+
+export interface Emotion {
+  css(template: TemplateStringsArray, ...args: Array<CSSInterpolation>): string
+  css(...args: Array<CSSInterpolation>): string
+  cx(...classNames: Array<ClassNamesArg>): string
+  flush(): void
+  hydrate(ids: Array<string>): void
+  injectGlobal(
+    template: TemplateStringsArray,
+    ...args: Array<CSSInterpolation>
+  ): void
+  injectGlobal(...args: Array<CSSInterpolation>): void
+  keyframes(
+    template: TemplateStringsArray,
+    ...args: Array<CSSInterpolation>
+  ): string
+  keyframes(...args: Array<CSSInterpolation>): string
+  sheet: CSSStyleSheet
+  cache: EmotionCache
+  merge(className: string): string
+  getRegisteredStyles(
+    registeredStyles: Array<string>,
+    className: string
+  ): string
+}
+
 function merge(
-  registered /*: Object */,
-  css /*: (*) => string */,
-  className /*: string */
+  registered: RegisteredCache,
+  css: Emotion['css'],
+  className: string
 ) {
-  const registeredStyles = []
+  const registeredStyles: string[] = []
   const rawClassName = getRegisteredStyles(
     registered,
     registeredStyles,
@@ -31,61 +87,29 @@ function merge(
   return rawClassName + css(registeredStyles)
 }
 
-/*
-export type Interpolation = any
-export type Interpolations = Array<Interpolation>
-
-type CreateStyles<ReturnValue> = (...args: Interpolations) => ReturnValue
-
-type ClassNameArg =
-  | string
-  | boolean
-  | { [key: string]: boolean | void | null }
-  | Array<ClassNameArg>
-  | void
-  | null
-
-declare class StyleSheet {
-  insert(rule: string): void;
-  flush(): void;
-  speedy(newVal: boolean): void;
-  tags: Array<HTMLStyleElement>;
-  isSpeedy: number;
-  ctr: number;
-}
-
-export type Emotion = {
-  css: CreateStyles<string>,
-  cx: (...classNames: Array<ClassNameArg>) => string,
-  flush: () => void,
-  hydrate: (ids: Array<string>) => void,
-  injectGlobal: CreateStyles<void>,
-  keyframes: CreateStyles<string>,
-  sheet: StyleSheet,
-  cache: EmotionCache,
-  merge,
-  getRegisteredStyles
-}
-*/
-
-let createEmotion = (options) /*: Emotion */ => {
+let createEmotion = (options: Options): Emotion => {
   let cache = createCache(options)
 
-  cache.sheet.speedy = function (value /*: boolean */) {
+  ;(cache.sheet as any).speedy = function (value: boolean) {
     if (process.env.NODE_ENV !== 'production' && this.ctr !== 0) {
       throw new Error('speedy must be changed before any rules are inserted')
     }
     this.isSpeedy = value
   }
+
   cache.compat = true
 
-  let css = (...args) => {
+  let css: Emotion['css'] = (
+    ...args: (TemplateStringsArray | Interpolation<unknown>)[]
+  ) => {
     let serialized = serializeStyles(args, cache.registered, undefined)
     insertStyles(cache, serialized, false)
     return `${cache.key}-${serialized.name}`
   }
 
-  let keyframes = (...args) => {
+  let keyframes: Emotion['keyframes'] = (
+    ...args: (TemplateStringsArray | Interpolation<unknown>)[]
+  ) => {
     let serialized = serializeStyles(args, cache.registered)
     let animation = `animation-${serialized.name}`
     insertWithoutScoping(cache, {
@@ -95,12 +119,14 @@ let createEmotion = (options) /*: Emotion */ => {
 
     return animation
   }
-  let injectGlobal = (...args) => {
+  let injectGlobal: Emotion['injectGlobal'] = (
+    ...args: (TemplateStringsArray | Interpolation<unknown>)[]
+  ) => {
     let serialized = serializeStyles(args, cache.registered)
     insertWithoutScoping(cache, serialized)
   }
 
-  let cx = (...args) => {
+  let cx: Emotion['cx'] = (...args) => {
     return merge(cache.registered, css, classnames(args))
   }
   return {
@@ -108,7 +134,7 @@ let createEmotion = (options) /*: Emotion */ => {
     cx,
     injectGlobal,
     keyframes,
-    hydrate(ids /*: Array<string> */) {
+    hydrate(ids) {
       ids.forEach(key => {
         cache.inserted[key] = true
       })
@@ -118,14 +144,14 @@ let createEmotion = (options) /*: Emotion */ => {
       cache.inserted = {}
       cache.sheet.flush()
     },
-    sheet: cache.sheet,
+    sheet: cache.sheet as CSSStyleSheet,
     cache,
     getRegisteredStyles: getRegisteredStyles.bind(null, cache.registered),
     merge: merge.bind(null, cache.registered, css)
   }
 }
 
-let classnames = args => {
+let classnames = (args: ClassNamesArg[]) => {
   let cls = ''
   for (let i = 0; i < args.length; i++) {
     let arg = args[i]
