+++ packages/css/types/index.d.ts
@@ -1,26 +1,1 @@
-// Definitions by: Junyoung Clare Jang <https://github.com/Ailrun>
-// TypeScript Version: 2.8
-
-import { Emotion } from './create-instance'
-
-export {
-  ArrayClassNamesArg,
-  ArrayCSSInterpolation,
-  ClassNamesArg,
-  ComponentSelector,
-  EmotionCache,
-  CSSInterpolation,
-  CSSObject,
-  CSSStyleSheet
-} from './create-instance'
-
-export const flush: Emotion['flush']
-export const hydrate: Emotion['hydrate']
-export const cx: Emotion['cx']
-export const merge: Emotion['merge']
-export const getRegisteredStyles: Emotion['getRegisteredStyles']
-export const css: Emotion['css']
-export const injectGlobal: Emotion['injectGlobal']
-export const keyframes: Emotion['keyframes']
-export const sheet: Emotion['sheet']
-export const cache: Emotion['cache']
+export * from '..'
